# PeaRL_spec/openapi/pearl-api-v1.1.yaml
openapi: 3.1.0
info:
  title: PeaRL API
  version: 1.1.0
  description: API-first risk orchestration platform for autonomous coding and secure/responsible
    delivery.
servers:
- url: http://localhost:8080
  description: Local dev
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ProjectId:
      name: project_id
      in: path
      required: true
      schema:
        type: string
        pattern: ^proj_[A-Za-z0-9_-]+$
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        minLength: 8
        maxLength: 200
    TraceId:
      name: X-Trace-Id
      in: header
      required: false
      schema:
        type: string
        minLength: 8
        maxLength: 128
  schemas:
...
    OrgBaseline:
      $ref: ../schemas/project/org-baseline.schema.json
    ApplicationSpec:
      $ref: ../schemas/project/application-spec.schema.json
    EnvironmentProfile:
      $ref: ../schemas/project/environment-profile.schema.json
...
    RemediationSpec:
      $ref: ../schemas/findings/remediation-spec.schema.json
...
    ErrorResponse:
      $ref: ../schemas/common/error-response.schema.json
  responses:
    BadRequest:
      description: Validation or contract error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
security:
- bearerAuth: []
paths:
...
  /api/v1/projects/{project_id}/org-baseline:
    post:
      tags:
      - Projects
      summary: Attach or update org baseline for project
      operationId: upsertOrgBaselineForProject
      parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/TraceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrgBaseline'
            examples:
              default:
                externalValue: ../examples/project/org-baseline.request.json
      responses:
        '200':
          description: Org baseline stored
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrgBaseline'
  /api/v1/projects/{project_id}/app-spec:
    post:
      tags:
      - Projects
      summary: Attach or update application spec
      operationId: upsertApplicationSpec
      parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/TraceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationSpec'
            examples:
              appSpec:
                externalValue: ../examples/project/app-spec.request.json
      responses:
        '200':
          description: Application spec stored
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationSpec'
  /api/v1/projects/{project_id}/environment-profile:
    post:
      tags:
      - Projects
      summary: Attach or update environment profile
      operationId: upsertEnvironmentProfile
      parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/TraceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnvironmentProfile'
            examples:
              envProfile:
                externalValue: ../examples/project/environment-profile.request.json
      responses:
        '200':
          description: Environment profile stored
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvironmentProfile'
...
  /api/v1/projects/{project_id}/remediation-specs/generate:
    post:
      tags:
      - Remediation
      summary: Generate remediation spec from normalized findings
      operationId: generateRemediationSpec
      parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/IdempotencyKey'
      - $ref: '#/components/parameters/TraceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                schema_version:
                  type: string
                finding_refs:
                  type: array
                  items:
                    type: string
                  minItems: 1
                environment:
                  type: string
                  enum:
                  - sandbox
                  - dev
                  - pilot
                  - preprod
                  - prod
                generation_mode:
                  type: string
                  enum:
                  - deterministic
                  - best_effort
                trace_id:
                  type: string
              required:
              - schema_version
              - finding_refs
              - environment
              - trace_id
            examples:
              remGen:
                externalValue: ../examples/remediation/generate-remediation-spec.request.json
      responses:
        '201':
          description: Remediation spec generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RemediationSpec'
              examples:
                remSpec:
                  externalValue: ../examples/remediation/generate-remediation-spec.response.json
        '400':
          $ref: '#/components/responses/BadRequest'
...